Add vsf-bridge compatibility
To workaround problems if vsf-bridge and magento1_external_checkout were
used (overriding of the same config.namespace) we gave it a new one.
We also added the usage of the "Secret" that is defined in vsf-bridge
instead of a new one.
Also the code pool was changed to be consistent with the modules
desciption file.