Force redirect and prevent the requested action from being executed
This PR solves #7 by throwing a `Mage_Core_Controller_Varien_Exception` in the observer's `preDispatch` method. This way the requested action is not executed anymore by Magento.
The background is that in the mentioned action the status code could be overwritten (e.g. in th noroute action), which makes the redirect useless.